# api/openapi/openapi.yaml
openapi: 3.0.3
info:
  title: kartochki.online API
  version: 0.1.0
  description: |
    Публичный API сервиса kartochki.online.

    Эта спецификация является источником правды для автогенерации frontend-клиента.
    Реализация backend и этот файл должны оставаться синхронными.
servers:
  - url: http://localhost:8080
paths:
...
  /api/v1/auth/telegram/login:
    post:
      summary: Вход через Telegram
      description: |
        Завершает вход через Telegram Login Widget.

        Frontend сначала показывает официальный Telegram Login Widget. После успешного входа
        Telegram возвращает набор подписанных полей: `id`, `first_name`, `last_name`,
        `username`, `photo_url`, `auth_date`, `hash`.

        Frontend должен отправить эти поля на backend без изменений. Backend проверяет подпись
        по токену Telegram-бота, убеждается что данные ещё не устарели, а затем находит
        или создаёт локального пользователя и открывает обычную локальную сессию.

        У Telegram может не быть email. Поэтому в ответе поле `user.email` может отсутствовать.
      operationId: loginWithTelegram
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TelegramLoginRequest'
      responses:
        '200':
          description: Вход через Telegram завершён, локальная сессия создана
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Подписанные данные Telegram отсутствуют, испорчены или устарели
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Вход через Telegram пока не настроен
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /api/v1/auth/forgot-password:
    post:
      summary: Запросить сброс пароля
      description: |
        Отправляет письмо со ссылкой для сброса пароля на указанный email.

        Ответ всегда `200` независимо от того, найден пользователь или нет.
        Это сделано намеренно, чтобы не раскрывать факт существования аккаунта по email.

        Все предыдущие активные токены сброса для этого пользователя инвалидируются —
        в каждый момент времени действует только один токен.
      operationId: forgotPassword
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequest'
      responses:
        '200':
          description: Запрос принят. Если аккаунт найден — письмо будет отправлено.
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Ошибка валидации входных данных
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/reset-password:
    post:
      summary: Установить новый пароль по токену из письма
      description: |
        Проверяет одноразовый токен из письма и устанавливает новый пароль.

        Токен действителен однократно и только в течение срока жизни, заданного на backend.
        После использования токен помечается и повторный запрос с тем же токеном вернёт ошибку.
      operationId: resetPassword
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequest'
      responses:
        '200':
          description: Пароль успешно изменён
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'
        '400':
          description: Токен недействителен, истёк или уже был использован; либо пароль не прошёл валидацию
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/vk/start:
    get:
      summary: Начать вход через VK
      description: |
        Первый шаг входа через VK ID.

        Frontend открывает этот endpoint, когда пользователь нажимает кнопку
        "Войти через VK". Backend подготавливает OAuth-запрос, сохраняет
        одноразовый `state` и перенаправляет пользователя на страницу VK ID.

        Этот endpoint не логинит пользователя сам по себе.
        Его задача только начать внешний OAuth-flow и отправить пользователя к провайдеру.
      operationId: startVkOAuth
      tags:
        - Auth
      responses:
        '302':
          description: Backend перенаправил пользователя на страницу VK ID для подтверждения входа
          headers:
            Location:
              description: URL страницы VK ID, куда нужно перейти для продолжения входа.
              schema:
                type: string
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
        '500':
          description: Не удалось начать OAuth-flow из-за внутренней ошибки backend
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Вход через VK пока не настроен
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/vk/callback:
    get:
      summary: Завершить вход через VK
      description: |
        Второй шаг входа через VK ID.

        Сюда VK ID вернёт пользователя после подтверждения входа.
        Обычно frontend не вызывает этот endpoint вручную.

        Backend проверяет `state`, получает профиль VK ID, находит или создаёт
        локального пользователя и затем открывает обычную локальную сессию backend.
      operationId: finishVkOAuth
      tags:
        - Auth
      parameters:
        - in: query
          name: code
          required: true
          description: Временный код авторизации, который VK ID выдаёт backend после успешного подтверждения входа.
          schema:
            type: string
        - in: query
          name: state
          required: true
          description: Защитный параметр, который помогает убедиться, что callback относится к нашему OAuth-запросу.
          schema:
            type: string
      responses:
        '200':
          description: Вход через VK завершён, локальная сессия создана
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Параметры `code` или `state` отсутствуют, истекли или не прошли проверку
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Не удалось завершить OAuth-flow из-за внутренней ошибки backend
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Вход через VK пока не настроен
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/yandex/start:
    get:
      summary: Начать вход через Яндекс
      description: |
        Первый шаг входа через Яндекс ID.

        Frontend открывает этот endpoint, когда пользователь нажимает кнопку
        "Войти через Яндекс". Backend подготавливает OAuth-запрос, сохраняет
        одноразовый `state` и перенаправляет пользователя на страницу Яндекса.
      operationId: startYandexOAuth
      tags:
        - Auth
      responses:
        '302':
          description: Backend перенаправил пользователя на страницу Яндекса для подтверждения входа
          headers:
            Location:
              description: URL страницы Яндекса, куда нужно перейти для продолжения входа.
              schema:
                type: string
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
        '500':
          description: Не удалось начать OAuth-flow из-за внутренней ошибки backend
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Вход через Яндекс пока не настроен
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/auth/yandex/callback:
    get:
      summary: Завершить вход через Яндекс
      description: |
        Второй шаг входа через Яндекс ID.

        Сюда Яндекс вернёт пользователя после подтверждения входа.
        Backend проверяет `state`, получает профиль Яндекс ID, находит или создаёт
        локального пользователя и затем открывает обычную локальную сессию backend.
      operationId: finishYandexOAuth
      tags:
        - Auth
      parameters:
        - in: query
          name: code
          required: true
          description: Временный код авторизации, который Яндекс выдаёт backend после успешного подтверждения входа.
          schema:
            type: string
        - in: query
          name: state
          required: true
          description: Защитный параметр, который помогает убедиться, что callback относится к нашему OAuth-запросу.
          schema:
            type: string
      responses:
        '200':
          description: Вход через Яндекс завершён, локальная сессия создана
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Параметры `code` или `state` отсутствуют, истекли или не прошли проверку
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Не удалось завершить OAuth-flow из-за внутренней ошибки backend
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '501':
          description: Вход через Яндекс пока не настроен
          headers:
            X-Request-ID:
              description: Идентификатор запроса для поиска в логах.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: 'Передавайте токен в заголовке Authorization: Bearer <token>.'
  schemas:
...
    StatusResponse:
      type: object
      required:
        - status
      properties:
        status:
          type: string
          description: Результат операции.
          enum:
            - accepted
            - password_changed
            - logged_out
            - deleted
          example: accepted
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: not_found
        message:
          type: string
          example: ресурс не найден
        request_id:
          type: string
          example: 01HZY3GQK0M5YAWQ2N6R8T9P1S
        details:
          type: array
          items:
            $ref: '#/components/schemas/ErrorDetail'
    ErrorDetail:
      type: object
      required:
        - message
      properties:
        field:
          type: string
          example: email
        message:
          type: string
          example: поле обязательно
...
    TelegramLoginRequest:
      type: object
      required:
        - id
        - auth_date
        - hash
      properties:
        id:
          type: integer
          format: int64
          example: 123456789
        first_name:
          type: string
          example: Иван
        last_name:
          type: string
          example: Иванов
        username:
          type: string
          example: ivanov
        photo_url:
          type: string
          format: uri
          example: https://t.me/i/userpic/320/example.jpg
        auth_date:
          type: integer
          format: int64
          example: 1712500000
        hash:
          type: string
          example: 7d3f6f8d3df0f70e18b1aabf6b8d2d11fdc4d4ab6f8d9c0d1c2b3a4f5e6d7c8b
    ForgotPasswordRequest:
      type: object
      required:
        - email
      properties:
        email:
          type: string
          format: email
          example: ivan@example.com
    ResetPasswordRequest:
      type: object
      required:
        - token
        - password
      properties:
        token:
          type: string
          description: Одноразовый токен из письма для сброса пароля.
          example: xK9mQ2rTv8nYpL3wZd5jF1cB6oE4sA7h
        password:
          type: string
          format: password
          description: Новый пароль пользователя.
          example: novyy-nadezhnyy-parol-456
    AuthUser:
      type: object
      required:
        - id
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: Иван
        email:
          type: string
          format: email
          nullable: true
          example: ivan@example.com
    AuthSession:
      type: object
      required:
        - access_token
        - token_type
        - expires_at
      properties:
        access_token:
          type: string
          example: 7fD2D1Xnq6C7xBskjPq6WQ6OHvOjQJE9P0Q-7m4w1RU
        token_type:
          type: string
          example: Bearer
        expires_at:
          type: string
          format: date-time
    AuthResponse:
      type: object
      required:
        - user
        - session
      properties:
        user:
          $ref: '#/components/schemas/AuthUser'
        session:
          $ref: '#/components/schemas/AuthSession'
...